+++ lib/honeybadger/agent.rb
@@ -214,8 +214,12 @@ module Honeybadger
     # @return [Boolean] true if the check in was successful and false
     #   otherwise.
     def check_in(id)
-      # this is to allow check ins even if a url is passed
-      check_in_id = id.to_s.strip.gsub(/\/$/, "").split("/").last
+      id_str = id.to_s.strip.gsub(/\/$/, "")
+      check_in_id = if id_str.include?("/check_in/")
+        id_str.sub(%r{.*/check_in/}, "")
+      else
+        id_str.split("/").last
+      end
       response = backend.check_in(check_in_id)
       response.success?
     end
@@ -422,6 +426,7 @@ module Honeybadger
         request_id = execution_context_manager.get_context[:request_id]
         p[:request_id] = request_id if request_id
         p[:hostname] = config[:hostname].to_s if config[:"events.attach_hostname"]
+        p[:environment] = config[:env].to_s if config[:"events.attach_environment"] && config[:env].to_s.length > 0
         p.update(event_context_manager.get_context)
       end