+++ lib/honeybadger/plugins/sidekiq.rb
@@ -56,6 +56,7 @@ module Honeybadger
 
         def call(worker, msg, queue, _redis)
           context = {
+            jid: msg["jid"],
             worker: msg["wrapped"] || msg["class"],
             queue: queue
           }
@@ -83,18 +84,8 @@ module Honeybadger
 
             if defined?(::Sidekiq::VERSION) && ::Sidekiq::VERSION > "3"
               ::Sidekiq.configure_server do |sidekiq|
-                sidekiq_default_configuration = (::Sidekiq::VERSION > "7") ?
-                  ::Sidekiq.default_configuration : Class.new
-
-                sidekiq.error_handlers << lambda { |ex, sidekiq_params, sidekiq_config = sidekiq_default_configuration|
+                sidekiq.error_handlers << lambda { |ex, sidekiq_params, _sidekiq_config = nil|
                   params = sidekiq_params.dup
-                  if defined?(::Sidekiq::Config)
-                    params[:_config] = if params[:_config].is_a?(::Sidekiq::Config) # Sidekiq > 6 and < 7.1.5
-                      params[:_config].instance_variable_get(:@options)
-                    else # Sidekiq >= 7.1.5
-                      sidekiq_config.instance_variable_get(:@options)
-                    end
-                  end
 
                   job = params[:job] || params
 
@@ -142,6 +133,7 @@ module Honeybadger
 
                 class SidekiqClusterCollectionChecker
                   include ::Sidekiq::Component
+
                   def initialize(config)
                     @config = config
                   end
@@ -222,7 +214,7 @@ module Honeybadger
         end
 
         collect do
-          if config.cluster_collection?(:sidekiq) && (leader_checker.nil? || leader_checker.collect?)
+          if ::Sidekiq.server? && config.cluster_collection?(:sidekiq) && (leader_checker.nil? || leader_checker.collect?)
             stats = collect_sidekiq_stats.call
 
             if Honeybadger.config.load_plugin_insights?(:sidekiq, feature: :events)